third_party/WebKit/LayoutTests/http/tests/navigation/same-document-scroll-position-restore.html

<body>
<div style="height: 1000px"></div>
<a id="a" href="#bottom">Link to bottom</a>
<div style="height: 1000px"></div>
<a name="bottom"></a>
<script src="/js-test-resources/js-test.js"></script>
<script>
setPrintTestResultsLazily();

if (window.testRunner) {
    window.jsTestIsAsync = true;
    testRunner.dumpAsText();
    testRunner.waitUntilDone();
}

window.onload = function() {
    if (window.eventSender) {
        function clickLink() {
            var a = document.getElementById("a");
            eventSender.mouseMoveTo(a.offsetLeft + 2, a.offsetTop - 798);
            eventSender.mouseDown();
            eventSender.mouseUp();
        }
        eventSender.continuousMouseScrollBy(0, -800);
        shouldBecomeEqual("window.scrollY == 800", "true", clickLink);
    }
};

var haveSeenHashChange =  false;

window.onhashchange = function() {
    if (!haveSeenHashChange) {
        haveSeenHashChange = true;
        history.back();
    } else {
        document.body.appendChild(document.createTextNode("scrollY should be 800: " + (window.scrollY == 800 ? "PASS" : "FAIL")));
        finishJSTest();
    }
}
</script>
</body>
